requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
distlib==0.3.7
Django==4.2.8
djangorestframework==3.14.0
filelock==3.12.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
install==1.3.5
msgpack==1.0.7
pbr==5.11.1
platformdirs==3.11.0
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
pytz==2023.3.post1
redis==5.0.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
stevedore==5.1.0
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
virtualenv==20.24.5
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
zope.interface==6.1
